refactor(OpportunityForm): replace old component with updated OpportunityForm and adjust imports in UpdateOpportunityModal
